requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3.post1
dateparser==1.2.0
pytest==7.4.3
gunicorn==21.2.0
